Add LocationController for location management
The location controller encapsulates the CLLocationManager functionalities
This controller may look weird due to the use of a generic parameter. The parameter is needed
as it made it possible to mock the class methods such as authorizationStatus() and locationServicesEnabled.
Both class methods are stubbed using a singleton while a CLLocationManager has been mocked in the tests